scripts/oltp/requirements.txt
psycopg2-binary>=2.9.0
pyyaml>=6.0